+++ packages/sdk/src/swap/sdk.ts
@@ -234,9 +234,6 @@ export class SwapSDK {
         }
       : undefined;
 
-    // DEPRECATED(1.5): use ccmParams instead of ccmMetadata
-    depositAddressRequest.ccmParams ??= depositAddressRequest.ccmMetadata; // eslint-disable-line no-param-reassign
-
     let response;
 
     if (this.options.broker !== undefined) {
@@ -314,9 +311,6 @@ export class SwapSDK {
     assert(signer, 'No signer provided');
 
     await this.validateSwapAmount({ chain: srcChain, asset: srcAsset }, BigInt(amount));
-
-    // DEPRECATED(1.5): use ccmParams instead of ccmMetadata
-    params.ccmParams ??= params.ccmMetadata; // eslint-disable-line no-param-reassign
 
     const tx = await executeSwap(
       params,