+++ packages/shared/src/vault/index.ts
@@ -13,9 +13,6 @@ export interface ExecuteSwapParams {
   amount: string;
   destAddress: string;
   ccmParams?: CcmParams;
-
-  /** @deprecated DEPRECATED(1.5): use ccmParams instead of ccmMetadata */
-  ccmMetadata?: CcmParams;
 }
 
 export type SwapNetworkOptions =